Added support for SVD mode. It's auto detected by checking whether vlan_tunnel is on for a VXLAN interface (there should be only one).
To use SVD mode, there should be a vlan-aware bridge, a vlan-aware
vxlan device associated with it, a veth with one side associated with
the vlan aware bridge and the other side should be the interface
specified in the zone setup.

Add a veth for guest network:
ip link add ethgsta type veth peer name ethgstb
ip link set dev ethgsta addrgenmode none # not necessary but safer and cleaner
ip link set dev ethgstb addrgenmode none # not necessary but safer and cleaner
ip link set dev ethgsta up
ip link set dev ethgstb up
ip link set dev ethgsta master br0

The ethgst<a|b> link is a trunk to the vlan aware bridge, the script
will create ethgstb.<vlanid>, add the tag to the bridge and allow it
to ethgsta and map the vlan to the corresponding vni on the vxlan
interface.

Signed-off-by: Julius <julius@1984.is>